finalbid_analyser/requirements.txt

streamlit
pandas
openpyxl


openai
langgraph
sentence-transformers

python-doctr[torch]
python-docx

python-dotenv
torch
torchvision